refactor: 공통 컴포넌트 추출
